Rename ord_schema.parquet_dataset -> ord_schema.parquet
With the new ord_schema.datasets module owning the high-level,
format-dispatching dataset I/O, the lower-level Parquet backend reads
cleaner as ord_schema.parquet (and datasets + parquet_dataset was
an awkward overlap). Renames the module and its test, updates all internal
imports, and leaves a back-compat ord_schema/parquet_dataset.py shim that
forwards attribute access to the new module with a DeprecationWarning
(removed in a future minor release). The pb_to_parquet_dataset script is
unaffected (different name).
